# internal/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Cost Deck Operator API
  description: |
    REST API for the Cost Deck Kubernetes FinOps operator.
    Provides namespace resource insights, workload scaling, optimization, and cluster monitoring.

    **Authentication:** All endpoints (except `/api/login`) require a valid `costdeck-session` cookie.
    Obtain one via `POST /api/login`.
  version: "1.3.0" # x-release-please-version
  contact:
    name: Cost Deck
    url: https://github.com/migalsp/costdeck-operator
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /
    description: Current instance
...
paths:
  /api/login:
    post:
      tags: [Auth]
      summary: Sign in
      description: Authenticate with username/password. Returns a session cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: costdeck-admin
                password:
                  type: string
                  example: "<from-secret>"
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: costdeck-session=eyJ...; Path=/; Max-Age=43200; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/logout:
    post:
      tags: [Auth]
      summary: Sign out
      description: Clears the session cookie.
      responses:
        "200":
          description: Logout successful
        "429":
          description: Too many failed sign-in attempts from this address
...
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Current user
      responses:
        "200":
          description: The signed-in identity
          content:
            application/json:
              schema:
                type: object
                properties:
                  sub: { type: string, example: "entra:7f3c..." }
                  name: { type: string }
                  email: { type: string }
                  role: { type: string, enum: [viewer, operator, admin] }
                  provider: { type: string, enum: [local, entra, anonymous] }
        "401":
          description: Not signed in
  /api/auth/entra/login:
    get:
      tags: [Auth]
      summary: Start Microsoft Entra ID sign-in
      description: Redirects the browser to Entra ID (authorization code flow with PKCE).
      parameters:
        - name: return
          in: query
          description: Path to return to after signing in.
          schema: { type: string, example: /scaling }
      responses:
        "302":
          description: Redirect to login.microsoftonline.com
...
  /api/auth/entra/callback-spa:
    post:
      tags: [Auth]
      summary: Complete Entra ID sign-in (single-page-app mode)
      description: Used by the /auth/callback dashboard route when the app registration uses the SPA redirect URI.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                code: { type: string }
                state: { type: string }
      responses:
        "200":
          description: Signed in; the session cookie is set
        "401":
          description: Sign-in rejected
...
  /api/cluster-info:
    get:
      tags: [System]
      summary: Cluster summary
      description: Returns node count and total cluster resources.
      responses:
        "200":
          description: Cluster info
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: integer
                  totalCPU:
                    type: string
                  totalMemory:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/cluster/nodes:
    get:
      tags: [System]
      summary: Node metrics
      description: Per-node resource metrics for the cluster heatmap.
      responses:
        "200":
          description: Node list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NodeMetrics"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/operator/logs:
    get:
      tags: [Health]
      summary: Operator logs
      description: Returns the trailing 100 lines of operator logs.
      responses:
        "200":
          description: Log text
          content:
            text/plain:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/namespaces:
    get:
      tags: [Namespaces]
      summary: List namespaces
      description: Lists all monitored NamespaceFinOps CRDs.
      responses:
        "200":
          description: Namespace list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamespaceFinOps"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/namespaces/{ns}/history:
    get:
      tags: [Namespaces]
      summary: Usage history
      description: Resource usage history for the last 60 minutes.
      parameters:
        - $ref: "#/components/parameters/Namespace"
      responses:
        "200":
          description: History data points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryPoint"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/namespaces/{ns}/pods:
    get:
      tags: [Namespaces]
      summary: Pod metrics
      description: Pod-level resource metrics for the namespace.
      parameters:
        - $ref: "#/components/parameters/Namespace"
      responses:
        "200":
          description: Pod list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PodDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/namespaces/{ns}/workloads:
    get:
      tags: [Namespaces]
      summary: List workloads
      description: Lists Deployments and StatefulSets in the namespace.
      parameters:
        - $ref: "#/components/parameters/Namespace"
      responses:
        "200":
          description: Workload list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkloadDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/namespaces/{ns}/workloads/{name}:
    put:
      tags: [Namespaces]
      summary: Scale workload
      description: Set replica count for a specific Deployment or StatefulSet.
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                kind:
                  type: string
                  enum: [Deployment, StatefulSet]
                replicas:
                  type: integer
      responses:
        "200":
          description: Scaled successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/namespaces/{ns}/optimize:
    post:
      tags: [Optimization]
      summary: Optimize namespace (deprecated)
      deprecated: true
      description: >-
        Rewrites the requests and limits of every workload from observed usage and stores the
        original values for revert. Deprecated: the dashboard and the assistant no longer
        change requests. Use GET /api/namespaces/{ns}/recommendations and apply the advice
        through your own deployment pipeline. Will be removed in a future major release.
      parameters:
        - $ref: "#/components/parameters/Namespace"
      responses:
        "200":
          description: Optimization applied
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/namespaces/{ns}/revert:
    post:
      tags: [Optimization]
      summary: Revert optimization
      description: Restore original resource requests/limits from before optimization.
      parameters:
        - $ref: "#/components/parameters/Namespace"
      responses:
        "200":
          description: Reverted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/namespaces/{ns}/recommendations:
    get:
      tags: [Optimization]
      summary: Right-sizing recommendations
      description: >-
        Compares each container's requests with observed demand: p95 CPU and peak memory over
        up to 14 days when VictoriaMetrics is connected, otherwise a single metrics-server
        reading with a larger margin. Read-only; results are cached for five minutes.
      parameters:
        - $ref: "#/components/parameters/Namespace"
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendations"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          description: No usage data is available for the namespace.
  /api/namespaces/{ns}/optimization:
    get:
      tags: [Optimization]
      summary: Optimization status
      description: Returns whether an optimization is active and the before/after values.
      parameters:
        - $ref: "#/components/parameters/Namespace"
      responses:
        "200":
          description: Optimization info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OptimizationStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/scaling/groups:
    get:
      tags: [Scaling]
      summary: List scaling groups
      responses:
        "200":
          description: Group list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScalingGroup"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Scaling]
      summary: Create scaling group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScalingGroup"
      responses:
        "201":
          description: Group created
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/scaling/groups/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Scaling]
      summary: Get scaling group
      responses:
        "200":
          description: Group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScalingGroup"
    put:
      tags: [Scaling]
      summary: Update scaling group
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScalingGroup"
      responses:
        "200":
          description: Group updated
    delete:
      tags: [Scaling]
      summary: Delete scaling group
      responses:
        "204":
          description: Group deleted
  /api/scaling/groups/{name}/manual:
    post:
      tags: [Scaling]
      summary: Manual override for group
      description: Activate or deactivate a scaling group manually.
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                active:
                  type: boolean
                  nullable: true
                  description: >
                    true forces scale up, false forces scale down, null clears the override
                    and hands control back to the schedule.
                activeUntil:
                  type: string
                  format: date-time
                  description: >
                    Optional deadline. The override is honoured until this timestamp, after
                    which the schedule takes over again. Must be in the future.
                until:
                  type: string
                  example: nextTransition
                  description: >
                    Relative alternative to activeUntil: "nextTransition" holds the override
                    until the schedule would change state on its own, or a Go duration such
                    as "4h". Clearing the override (active null) also removes the legacy
                    costdeck.io/manual-override annotation.
      responses:
        "200":
          description: Override applied
        "400":
          description: Invalid deadline or duration
  /api/scaling/configs:
    get:
      tags: [Scaling]
      summary: List scaling configs
      responses:
        "200":
          description: Config list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScalingConfig"
    post:
      tags: [Scaling]
      summary: Create scaling config
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScalingConfig"
      responses:
        "201":
          description: Config created
  /api/scaling/configs/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Scaling]
      summary: Get scaling config
      responses:
        "200":
          description: Config details
    put:
      tags: [Scaling]
      summary: Update scaling config
      responses:
        "200":
          description: Config updated
    delete:
      tags: [Scaling]
      summary: Delete scaling config
      responses:
        "204":
          description: Config deleted
  /api/scaling/configs/{name}/manual:
    post:
      tags: [Scaling]
      summary: Manual override for config
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                active:
                  type: boolean
                  nullable: true
                  description: >
                    true forces scale up, false forces scale down, null clears the override
                    and hands control back to the schedule.
                activeUntil:
                  type: string
                  format: date-time
                  description: >
                    Optional deadline. The override is honoured until this timestamp, after
                    which the schedule takes over again. Must be in the future.
                until:
                  type: string
                  example: nextTransition
                  description: >
                    Relative alternative to activeUntil: "nextTransition" holds the override
                    until the schedule would change state on its own, or a Go duration such
                    as "4h". Clearing the override (active null) also removes the legacy
                    costdeck.io/manual-override annotation.
      responses:
        "200":
          description: Override applied
        "400":
          description: Invalid deadline or duration
components:
  parameters:
    Namespace:
      name: ns
      in: path
      required: true
      description: Target namespace name
      schema:
        type: string
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Authentication required
    NodeMetrics:
      type: object
      properties:
        name:
          type: string
        cpuCapacity:
          type: string
        cpuUsage:
          type: string
        memoryCapacity:
          type: string
        memoryUsage:
          type: string
        pods:
          type: integer
...
    ResourceMetrics:
      type: object
      properties:
        usage:
          type: string
        requests:
          type: string
        limits:
          type: string
    HistoryPoint:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        cpu:
          $ref: "#/components/schemas/ResourceMetrics"
        memory:
          $ref: "#/components/schemas/ResourceMetrics"
    PodDetail:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        cpu:
          $ref: "#/components/schemas/ResourceMetrics"
        memory:
          $ref: "#/components/schemas/ResourceMetrics"
    WorkloadDetail:
      type: object
      properties:
        name:
          type: string
        kind:
          type: string
          enum: [Deployment, StatefulSet]
        replicas:
          type: integer
        readyReplicas:
          type: integer
        status:
          type: string
    NamespaceFinOps:
      type: object
      properties:
        metadata:
          type: object
          properties:
            name:
              type: string
            namespace:
              type: string
        spec:
          type: object
          properties:
            targetNamespace:
              type: string
        status:
          type: object
          properties:
            insights:
              type: array
              items:
                type: string
    Recommendations:
      type: object
      properties:
        namespace: { type: string }
        source: { type: string, enum: [victoriametrics, metrics-server] }
        historical: { type: boolean, description: "true for p95/peak over a window, false for a single reading" }
        window: { type: string, example: 14d }
        basis: { type: string }
        warning: { type: string }
        currency: { type: string, example: USD }
        monthlySavings: { type: number }
        workloads:
          type: array
          items:
            type: object
            properties:
              kind: { type: string, enum: [Deployment, StatefulSet] }
              name: { type: string }
              replicas: { type: integer }
              monthlySavings: { type: number }
              containers:
                type: array
                items:
                  type: object
                  properties:
                    name: { type: string }
                    cpu: { $ref: "#/components/schemas/ResourceAdvice" }
                    memory: { $ref: "#/components/schemas/ResourceAdvice" }
    ResourceAdvice:
      type: object
      properties:
        request: { type: string, example: "1" }
        observed: { type: string, example: 100m }
        recommended: { type: string, example: 120m }
        action: { type: string, enum: [reduce, increase, keep, set, unknown] }
    OptimizationStatus:
      type: object
      properties:
        active:
          type: boolean
        optimizedAt:
          type: string
          format: date-time
        workloads:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              kind:
                type: string
              original:
                $ref: "#/components/schemas/ResourceValues"
              optimized:
                $ref: "#/components/schemas/ResourceValues"
    ResourceValues:
      type: object
      properties:
        cpuRequest:
          type: string
        cpuLimit:
          type: string
        memoryRequest:
          type: string
        memoryLimit:
          type: string
    ScalingGroup:
      type: object
      properties:
        metadata:
          type: object
          properties:
            name:
              type: string
        spec:
          type: object
          properties:
            category:
              type: string
            namespaces:
              type: array
              items:
                type: string
            active:
              type: boolean
              nullable: true
              description: Manual override. Null (absent) means the schedule is followed.
            activeUntil:
              type: string
              format: date-time
              description: Optional expiry for the manual override.
            schedules:
              type: array
              items:
                $ref: "#/components/schemas/ScalingSchedule"
            sequence:
              type: array
              items:
                type: string
    ScalingConfig:
      type: object
      properties:
        metadata:
          type: object
          properties:
            name:
              type: string
        spec:
          type: object
          properties:
            targetNamespace:
              type: string
            active:
              type: boolean
              nullable: true
              description: Manual override. Null (absent) means the schedule is followed.
            activeUntil:
              type: string
              format: date-time
              description: Optional expiry for the manual override.
            schedules:
              type: array
              items:
                $ref: "#/components/schemas/ScalingSchedule"
    ScalingSchedule:
      type: object
      description: >
        One availability window. Either a daily window (days + startTime/endTime, where an
        endTime earlier than startTime runs overnight into the next day) or a continuous
        multi-day window (startDay/endDay), which may wrap through Sunday.
      properties:
        days:
          type: array
          description: Weekdays the daily window applies to (0 = Sunday). Ignored when startDay/endDay are set.
          items:
            type: integer
            minimum: 0
            maximum: 6
        startDay:
          type: integer
          minimum: 0
          maximum: 6
          description: Weekday a continuous window opens on. Must be set together with endDay.
          example: 1
        endDay:
          type: integer
          minimum: 0
          maximum: 6
          description: Weekday a continuous window closes on. Must be set together with startDay.
          example: 5
        startTime:
          type: string
          example: "08:00"
        endTime:
          type: string
          example: "20:00"
        timezone:
          type: string
          example: Europe/Bratislava